requirements.txt
picamera
Pillow
python-telegram-bot
